services/api/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
asyncpg
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-email-validator
fastapi-mail>=1.4.1
python-multipart
motor
beanie
httpx
apscheduler
pytest
requests
locust
